Shifting theming for better nextjs style alignment
