+++ app/components/ThemeWrapper.tsx
@@ -1,0 +1,15 @@
+'use client';
+import * as React from 'react';
+import { ThemeProvider } from '@mui/material/styles';
+import CssBaseline from '@mui/material/CssBaseline';
+import theme from './theme';
+
+export default function ThemeWrapper({ children }: { children: React.ReactNode }) {
+  return (
+      <ThemeProvider theme={theme}>
+        {/* CssBaseline kickstart an elegant, consistent, and simple baseline to build upon. */}
+        <CssBaseline />
+        {children}
+      </ThemeProvider>
+  );
+}
